NIFI-14921 Decoupled ZooKeeper to Separate NAR
- Added framework extension bundle and NAR for ZooKeeper
- Moved ZooKeeper Leader Election Manager and State Provider classes
- Added ZooKeeperStateProviderServer wrapper for embedded ZooKeeper
- Updated FlowController to load embedded ZooKeeper using Extension Manager with State Provider
