+++ nifi-assembly/pom.xml
@@ -833,6 +833,12 @@ language governing permissions and limitations under the License. -->
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
+            <artifactId>nifi-framework-zookeeper-nar</artifactId>
+            <version>2.6.0-SNAPSHOT</version>
+            <type>nar</type>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-compress-nar</artifactId>
             <version>2.6.0-SNAPSHOT</version>
             <type>nar</type>
