+++ nifi-framework-bundle/nifi-framework-extensions/nifi-framework-zookeeper-bundle/nifi-framework-cluster-zookeeper/pom.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
   Licensed to the Apache Software Foundation (ASF) under one or more
-  contributor license agreements. See the NOTICE file distributed with
+  contributor license agreements.  See the NOTICE file distributed with
   this work for additional information regarding copyright ownership.
   The ASF licenses this file to You under the Apache License, Version 2.0
   (the "License"); you may not use this file except in compliance with
-  the License. You may obtain a copy of the License at
-  http://www.apache.org/licenses/LICENSE-2.0
+  the License.  You may obtain a copy of the License at
+      http://www.apache.org/licenses/LICENSE-2.0
   Unless required by applicable law or agreed to in writing, software
   distributed under the License is distributed on an "AS IS" BASIS,
   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,20 +17,27 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.apache.nifi</groupId>
-        <artifactId>nifi-framework-bundle</artifactId>
+        <artifactId>nifi-framework-zookeeper-bundle</artifactId>
         <version>2.6.0-SNAPSHOT</version>
     </parent>
-    <artifactId>nifi-framework-extensions</artifactId>
-    <packaging>pom</packaging>
-    <modules>
-        <module>nifi-framework-kubernetes-bundle</module>
-        <module>nifi-framework-zookeeper-bundle</module>
-        <module>nifi-kerberos-iaa-providers-bundle</module>
-        <module>nifi-ldap-iaa-providers-bundle</module>
-        <module>nifi-provenance-repository-bundle</module>
-        <module>nifi-py4j-framework-bundle</module>
-        <module>nifi-questdb-bundle</module>
-        <module>nifi-single-user-iaa-providers-bundle</module>
-    </modules>
+    <artifactId>nifi-framework-cluster-zookeeper</artifactId>
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.nifi</groupId>
+            <artifactId>nifi-utils</artifactId>
+            <version>2.6.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.zookeeper</groupId>
+            <artifactId>zookeeper</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.curator</groupId>
+            <artifactId>curator-client</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-lang3</artifactId>
+        </dependency>
+    </dependencies>
 </project>
-
