+++ nifi-framework-bundle/nifi-framework-extensions/nifi-framework-zookeeper-bundle/nifi-framework-zookeeper-leader-election/src/main/resources/META-INF/services/org.apache.nifi.controller.leader.election.LeaderElectionManager
