+++ nifi-framework-bundle/nifi-framework-extensions/nifi-framework-zookeeper-bundle/nifi-framework-zookeeper-state-provider/src/main/resources/META-INF/services/org.apache.nifi.components.state.StateProvider
@@ -12,4 +12,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-org.apache.nifi.controller.state.providers.local.WriteAheadLocalStateProvider
+org.apache.nifi.controller.state.providers.zookeeper.ZooKeeperStateProvider
+org.apache.nifi.controller.state.providers.zookeeper.server.ZooKeeperStateProviderServer
+
