+++ nifi-framework-bundle/nifi-framework-nar-bom/pom.xml
@@ -95,12 +95,6 @@
             </dependency>
             <dependency>
                 <groupId>org.apache.nifi</groupId>
-                <artifactId>nifi-framework-zookeeper-leader-election</artifactId>
-                <version>2.6.0-SNAPSHOT</version>
-                <scope>provided</scope>
-            </dependency>
-            <dependency>
-                <groupId>org.apache.nifi</groupId>
                 <artifactId>nifi-framework-nar-utils</artifactId>
                 <version>2.6.0-SNAPSHOT</version>
                 <scope>provided</scope>
@@ -261,10 +255,6 @@
         <dependency>
             <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-utils</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-framework-zookeeper-leader-election</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
