--- nifi-framework-bundle/nifi-framework/nifi-framework-core/src/main/java/org/apache/nifi/controller/state/NodeDescription.java
@@ -1,39 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.nifi.controller.state;
-
-/**
- * Provides information about a node in a NiFi cluster
- */
-public interface NodeDescription {
-    /**
-     * @return the unique identifier for this node in the cluster
-     */
-    String getNodeIdentifier();
-
-    /**
-     * @return the hostname of the node
-     */
-    String getHostname();
-
-    /**
-     * @return the port on which the node's embedded ZooKeeper Server is running, or <code>null</code> if the node is
-     *         not running an embedded ZooKeeper server
-     */
-    Integer getEmbeddedZooKeeperPort();
-}
