+++ nifi-system-tests/nifi-system-test-suite/pom.xml
@@ -296,6 +296,12 @@
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
+            <artifactId>nifi-framework-zookeeper-nar</artifactId>
+            <version>2.6.0-SNAPSHOT</version>
+            <type>nar</type>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-server-nar</artifactId>
             <version>2.6.0-SNAPSHOT</version>
             <type>nar</type>
